renames test_efficientdet2 -> test_efficientdet
